Arrange state store failures in the test
